+++ services/tip.service.js
@@ -1,0 +1,12 @@
+const User = require("../schemas/user")
+
+exports.postipService = async (userId, dataId) => {
+    return await User.updateOne(
+        { userId },
+        {
+            $push: {
+                dismatchData: [dataId],
+            },
+        }
+    )
+}
